progress

+++ java_console/models/src/com/rusefi/autotune/Result.java
@@ -5,13 +5,13 @@ package com.rusefi.autotune;
  * 2/23/2016.
  */
 public class Result {
-    private final float[][] kgbcRES;
+    private final double[][] kgbcRES;
 
-    public Result(float[][] kgbcRES) {
+    public Result(double[][] kgbcRES) {
         this.kgbcRES = kgbcRES;
     }
 
-    public float[][] getKgbcRES() {
+    public double[][] getKgbcRES() {
         return kgbcRES;
     }
 }
